+++ CHANGELOG.md
@@ -78,6 +78,7 @@
 
 ### Behavior change
 - `Dataframe.columns` now returns native pandas Index object instead of Snowpark Index object.
+- Refactor and introduce `query_compiler` argument in `Index` constructor to create `Index` from query compiler.
 
 ## 1.20.0 (2024-07-17)